backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
openai
python-dotenv
alembic
requests
google-generativeai
pillow
